WIP: consolidate and refactor API
